+++ src/server/workers/lang/modules/documentation.ts
@@ -1,0 +1,68 @@
+import * as types from "../../../../common/types";
+import * as activeProject from "../activeProject";
+import * as utils from "../../../../common/utils";
+
+//--------------------------------------------------------------------------
+//  getNavigateToItems
+//--------------------------------------------------------------------------
+
+// Look at
+// https://github.com/Microsoft/TypeScript/blob/master/src/services/navigateTo.ts
+// for inspiration
+// Reason for forking:
+//  didn't give all results
+//  gave results from lib.d.ts
+//  I wanted the practice
+
+export function getNavigateToItems(query: {}): Promise<types.GetNavigateToItemsResponse> {
+    let project = activeProject.GetProject.getCurrentIfAny();
+    var languageService = project.languageService;
+
+    let getNodeKind = ts.getNodeKind;
+    function getDeclarationName(declaration: ts.Declaration): string {
+        let result = getTextOfIdentifierOrLiteral(declaration.name);
+        if (result !== undefined) {
+            return result;
+        }
+
+        if (declaration.name.kind === ts.SyntaxKind.ComputedPropertyName) {
+            let expr = (<ts.ComputedPropertyName>declaration.name).expression;
+            if (expr.kind === ts.SyntaxKind.PropertyAccessExpression) {
+                return (<ts.PropertyAccessExpression>expr).name.text;
+            }
+
+            return getTextOfIdentifierOrLiteral(expr);
+        }
+
+        return undefined;
+    }
+    function getTextOfIdentifierOrLiteral(node: ts.Node) {
+        if (node.kind === ts.SyntaxKind.Identifier ||
+            node.kind === ts.SyntaxKind.StringLiteral ||
+            node.kind === ts.SyntaxKind.NumericLiteral) {
+
+            return (<ts.Identifier | ts.LiteralExpression>node).text;
+        }
+
+        return undefined;
+    }
+
+    var items: types.NavigateToItem[] = [];
+    for (let file of project.getProjectSourceFiles()) {
+        let declarations = file.getNamedDeclarations();
+        for (let index in declarations) {
+            for (let declaration of declarations[index]) {
+                let item: types.NavigateToItem = {
+                    name: getDeclarationName(declaration),
+                    kind: getNodeKind(declaration),
+                    filePath: file.fileName,
+                    fileName: utils.getFileName(file.fileName),
+                    position: project.languageServiceHost.getLineAndCharacterOfPosition(file.fileName, declaration.getStart())
+                }
+                items.push(item);
+            }
+        }
+    }
+
+    return utils.resolve({ items });
+}
