+++ src/server/workers/lang/projectService.ts
@@ -285,6 +285,9 @@ export function getReferences(query: Types.GetReferencesQuery): Promise<Types.Ge
     })
 }
 
+/**
+ * Formatting
+ */
 import * as formatting from "./modules/formatting";
 export function formatDocument(query: Types.FormatDocumentQuery): Promise<Types.FormatDocumentResponse> {
     let project = getProject(query.filePath);
@@ -295,71 +298,11 @@ export function formatDocumentRange(query: Types.FormatDocumentRangeQuery): Prom
     return resolve({ refactorings: formatting.formatDocumentRange(project, query.filePath, query.from, query.to, query.editorOptions) });
 }
 
-
-//--------------------------------------------------------------------------
-//  getNavigateToItems
-//--------------------------------------------------------------------------
-
-// Look at
-// https://github.com/Microsoft/TypeScript/blob/master/src/services/navigateTo.ts
-// for inspiration
-// Reason for forking:
-//  didn't give all results
-//  gave results from lib.d.ts
-//  I wanted the practice
-
-export function getNavigateToItems(query: {}): Promise<Types.GetNavigateToItemsResponse> {
-    let project = activeProject.GetProject.getCurrentIfAny();
-    var languageService = project.languageService;
-
-    let getNodeKind = ts.getNodeKind;
-    function getDeclarationName(declaration: ts.Declaration): string {
-        let result = getTextOfIdentifierOrLiteral(declaration.name);
-        if (result !== undefined) {
-            return result;
-        }
-
-        if (declaration.name.kind === ts.SyntaxKind.ComputedPropertyName) {
-            let expr = (<ts.ComputedPropertyName>declaration.name).expression;
-            if (expr.kind === ts.SyntaxKind.PropertyAccessExpression) {
-                return (<ts.PropertyAccessExpression>expr).name.text;
-            }
-
-            return getTextOfIdentifierOrLiteral(expr);
-        }
-
-        return undefined;
-    }
-    function getTextOfIdentifierOrLiteral(node: ts.Node) {
-        if (node.kind === ts.SyntaxKind.Identifier ||
-            node.kind === ts.SyntaxKind.StringLiteral ||
-            node.kind === ts.SyntaxKind.NumericLiteral) {
-
-            return (<ts.Identifier | ts.LiteralExpression>node).text;
-        }
-
-        return undefined;
-    }
-
-    var items: Types.NavigateToItem[] = [];
-    for (let file of project.getProjectSourceFiles()) {
-        let declarations = file.getNamedDeclarations();
-        for (let index in declarations) {
-            for (let declaration of declarations[index]) {
-                let item: Types.NavigateToItem = {
-                    name: getDeclarationName(declaration),
-                    kind: getNodeKind(declaration),
-                    filePath: file.fileName,
-                    fileName: utils.getFileName(file.fileName),
-                    position: project.languageServiceHost.getLineAndCharacterOfPosition(file.fileName, declaration.getStart())
-                }
-                items.push(item);
-            }
-        }
-    }
-
-    return resolve({ items });
-}
+/**
+ * Documentation
+ */
+import * as documentation from "./modules/documentation";
+export const getNavigateToItems = documentation.getNavigateToItems;
 
 /**
  * Dependency View
